feat(delimiter): introduce new delimiter type NONE
this commit doesn't contain the actual delimiter change. The actual change was done in another PR by
a community member. The commit messages in the PR that contains the actual change did not follow
conventional commits and therefore couldn't be used to automatically release the feature. Hence this
proxy commit.